Move schedule into src/js
